requirements.txt
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
DateTime==4.3
idna==2.8
Pillow==6.2.1
pytz==2019.3
qrcode==6.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
zope.interface==4.7.1
